% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{dottools-package}
\alias{dottools}
\alias{dottools-package}
\title{dottools: Helper Functions for ANVEA-DOT}
\description{
What the package does (one paragraph).
}
\section{Package Options}{

\describe{
 \item{\code{dottools.filter.coverage}}{Minimum coverage for considering a
 gene in ANEVA-DOT}
 \item{\code{dottools.test.mono_cutoff}}{Other allele ratio for monoallelic
 test; see \code{\link{monoallelic_test}} for more details}
 \item{\code{dottools.test.mono_threshold}}{\emph{p}-value threshold for
 monoallelic test; see \code{\link{monoallelic_test}} for more details}
 \item{\code{dottools.test.homozygous_method}}{Method for running homozygous
 test; defaults to \dQuote{auto}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LappalainenLab/dottools}
  \item Report bugs at \url{https://github.com/LappalainenLab/dottools/issues}
}

}
\author{
\strong{Maintainer}: Paul Hoffman \email{phoffman@nygenome.org} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID})

}
