% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dottests.R
\name{homozygous_test}
\alias{homozygous_test}
\title{Perform the Homozygous Test}
\usage{
homozygous_test(
  ase,
  method = getOption(x = "dottools.test.homozygous_method"),
  missing = c("homozygous", "na"),
  other = TRUE,
  cl = NULL,
  verbose = TRUE
)
}
\arguments{
\item{ase}{An ASE data frame from \code{\link{read_ase}}}

\item{method}{Method for running the homozygous test; choose from
\itemize{
 \item \dQuote{\code{genotype}}: Use genotype information for determining
  heterozygous/homozygous status
 \item \dQuote{\code{counts}}: Use ref/alt counts to determine
  heterozygous/homozygous status
 \item \dQuote{\code{auto}}: If a \dQuote{\code{GENOTYPE}} column present,
  use the \dQuote{\code{genotype}} method; otherwise, use the
  \dQuote{\code{counts}} method
}}

\item{missing}{For \code{method == "genotype"}, behavior for missing
genotypes; choose from:
\itemize{
 \item \dQuote{\code{homozygous}}: Treat missing genotypes as homozygous
 \item \dQuote{\code{na}}: Mark missing genotypes with \code{NA}
}}

\item{other}{For \code{method == "counts"}, include other counts with alt
counts in heterozygous/homozygous test}

\item{cl}{
A cluster object created by \code{\link{makeCluster}},
or an integer to indicate number of child-processes
(integer values are ignored on Windows) for parallel evaluations
(see Details on performance).
}

\item{verbose}{Show progress updates}
}
\value{
\code{ase} with the following column added:
\itemize{
 \item \dQuote{\code{HOMOZYGOUS_WARNING}}: Flag to indicate if the variant
  associated with the gene is homozygous
}
}
\description{
Perform the Homozygous Test
}
\seealso{
Other ASE filtering tests:
\code{\link{bias_test}()},
\code{\link{mapability_test}()},
\code{\link{monoallelic_test}()}
}
\concept{filter_test}
