% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dottests.R
\name{monoallelic_test}
\alias{monoallelic_test}
\title{Perform the Monoallelic Test}
\usage{
monoallelic_test(
  ase,
  coverage = getOption(x = "dottools.filter.coverage"),
  cutoff = getOption(x = "dottools.test.mono_cutoff"),
  threshold = getOption(x = "dottools.test.mono_threshold"),
  verbose = TRUE
)
}
\arguments{
\item{ase}{An ASE data frame from \code{\link{read_ase}}}

\item{coverage}{...}

\item{cutoff}{Maximum proportion of reads arising from non ref/alt read for
variant to be counted}

\item{threshold}{\emph{p}-value threshold}

\item{verbose}{Show progress updates}
}
\value{
\code{ase} with the following columns added:
\itemize{
 \item \dQuote{\code{mono_pvals}}: \emph{p}-values for the monoallelic test
 \item \dQuote{\code{mono_qvals}}: FDR-adjusted \emph{p}-values
  (\emph{q}-values) for the monoallelic test
 \item \dQuote{\code{GENOTYPE_WARNING}}: Flag to indicate if the gene fails
  the monoallelic test
}
}
\description{
Perform the Monoallelic Test
}
\seealso{
Other ASE filtering tests:
\code{\link{bias_test}()},
\code{\link{homozygous_test}()},
\code{\link{mapability_test}()}
}
\concept{filter_test}
