% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{prep_vgs}
\alias{prep_vgs}
\title{Prepare a Vgs for ANEVA-DOT}
\usage{
prep_vgs(ase, vgs, verbose = TRUE)
}
\arguments{
\item{ase}{A data frame with ASE from \code{\link{read_ase}}}

\item{vgs}{A data frame from \code{\link{read_vgs}}}

\item{verbose}{Show progress updates}
}
\value{
A named vector with the Vgs to use; names are the gene IDs, values
are the Vgs
}
\description{
Generate a vector of Vgs for input into ANEVA-DOT
}
