% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_blacklist}
\alias{read_blacklist}
\title{Read in a Gene Blacklist}
\usage{
read_blacklist(blfile, verbose = TRUE)
}
\arguments{
\item{blfile}{Path to the gene blacklist}

\item{verbose}{Show progress updates}
}
\value{
A character vector of genes to exclude from analysis
}
\description{
Read in a Gene Blacklist
}
\concept{data_input}
