% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_symbol_map}
\alias{read_symbol_map}
\title{Read in a Gene Symbol Map}
\usage{
read_symbol_map(mapfile, verbose = TRUE)
}
\arguments{
\item{mapfile}{Path to the gene ID/symbol map file; should be a
tab-separated file with ...}

\item{verbose}{Show progress updates}
}
\value{
A \code{\link[data.table:data.table-class]{data.table}} with two
columns:
\itemize{
 \item \dQuote{\code{GENE_ID}} ...
 \item \dQuote{\code{GENE_SYMBOL}} ...
}
}
\description{
Read in a Gene Symbol Map
}
\concept{data_input}
